Subo el GetVertex() :)

+++ randomgame/src/Servidor/modelo/Physics/Bodies/Cuadrado.h
@@ -6,10 +6,10 @@
 
 class Cuadrado: public Body{
 public:
-	Cuadrado(ElementType type, int posX, int posY, float h, 
+	Cuadrado(ElementType type, float posX, float posY, float h, 
 		float w, float masa, float angle, b2World *myWorld, GameElement* modelElement);
 	virtual ~Cuadrado();
-	std::list<std::pair<int,int>> GetVertex();
+	std::list<std::pair<float,float>> GetVertex();
 	void animate();
 	void touch(Body* touchingWith, b2World* mundo);
